Refactoring of expressions.dart
