+++ packages/expression_language/lib/src/expressions/datetime_minus_duration.dart
@@ -1,0 +1,17 @@
+import 'package:expression_language/expression_language.dart';
+
+class DateTimeMinusDurationExpression extends Expression<DateTime> {
+  final Expression<DateTime> left;
+  final Expression<Duration> right;
+
+  DateTimeMinusDurationExpression(this.left, this.right);
+  @override
+  DateTime evaluate() {
+    return left.evaluate().subtract(right.evaluate());
+  }
+
+  @override
+  void accept(ExpressionVisitor visitor) {
+    visitor.visitDateTimeMinusDuration(this);
+  }
+}
