+++ packages/expression_language/lib/src/expressions/equal_number.dart
@@ -1,0 +1,18 @@
+import 'package:expression_language/expression_language.dart';
+
+class EqualNumberExpression extends Expression<bool> {
+  final Expression<Number> left;
+  final Expression<Number> right;
+
+  EqualNumberExpression(this.left, this.right);
+
+  @override
+  bool evaluate() {
+    return left.evaluate() == right.evaluate();
+  }
+
+  @override
+  void accept(ExpressionVisitor visitor) {
+    visitor.visitEqualNumber(this);
+  }
+}
