+++ packages/expression_language/lib/src/expressions/less_than_number.dart
@@ -1,0 +1,18 @@
+import 'package:expression_language/expression_language.dart';
+
+class LessThanNumberExpression extends Expression<bool> {
+  final Expression<Number> left;
+  final Expression<Number> right;
+
+  LessThanNumberExpression(this.left, this.right);
+
+  @override
+  bool evaluate() {
+    return left.evaluate() < right.evaluate();
+  }
+
+  @override
+  void accept(ExpressionVisitor visitor) {
+    visitor.visitLessThanNumber(this);
+  }
+}
