+++ packages/expression_language/lib/src/expressions/less_than_or_equal_duration.dart
@@ -1,0 +1,18 @@
+import 'package:expression_language/expression_language.dart';
+
+class LessThanOrEqualDurationExpression extends Expression<bool> {
+  final Expression<Duration> left;
+  final Expression<Duration> right;
+
+  LessThanOrEqualDurationExpression(this.left, this.right);
+
+  @override
+  bool evaluate() {
+    return left.evaluate() <= right.evaluate();
+  }
+
+  @override
+  void accept(ExpressionVisitor visitor) {
+    visitor.visitLessThanOrEqualDuration(this);
+  }
+}
