+++ packages/expression_language/lib/src/expressions/logical_and.dart
@@ -1,0 +1,18 @@
+import 'package:expression_language/expression_language.dart';
+
+class LogicalAndExpression extends Expression<bool> {
+  final Expression<bool> left;
+  final Expression<bool> right;
+
+  LogicalAndExpression(this.left, this.right);
+
+  @override
+  bool evaluate() {
+    return left.evaluate() && right.evaluate();
+  }
+
+  @override
+  void accept(ExpressionVisitor visitor) {
+    visitor.visitLogicalAnd(this);
+  }
+}
