+++ packages/expression_language/lib/src/expressions/minus_duration.dart
@@ -1,0 +1,18 @@
+import 'package:expression_language/expression_language.dart';
+
+class MinusDurationExpression extends Expression<Duration> {
+  final Expression<Duration> left;
+  final Expression<Duration> right;
+
+  MinusDurationExpression(this.left, this.right);
+
+  @override
+  Duration evaluate() {
+    return left.evaluate() - right.evaluate();
+  }
+
+  @override
+  void accept(ExpressionVisitor visitor) {
+    visitor.visitMinusDuration(this);
+  }
+}
